feat: Enhanced B2B Flutter app - bulk delete, settlement tracking, password auth, top products
- Added bulk product selection and deletion in admin products page with bottom action bar
- Implemented comprehensive settlement tracking in ledger (delivered/sold/pending items)
- Added password authentication for superuser login with dynamic UI and error handling
- Created Top Products carousel management page for home page featured products
- Added detailed product details page with full info, image, and add to cart
- Implemented Mark Sold functionality for subdealers with real-time validation
- Added Settlement Report page for tracking money settlements with progress indicators
- Enhanced catalog with product click-through navigation to details page
- Improved home page with Top Products carousel section
- Enhanced UI/UX across multiple pages with better error handling and visual feedback
- Added settlement fields to UserBalance and OrderItem models
- Fixed cart quantity handling with proper double conversion
- All features tested and validated with no compilation errors

+++ lib/features/orders/checkout_page.dart
@@ -131,7 +131,7 @@ class _CheckoutPageState extends State<CheckoutPage> {
 
       // Place order
       final items = cartQty.value.entries
-          .map((e) => {"product_id": e.key, "qty": e.value})
+          .map((e) => {"product_id": e.key, "qty": e.value.toDouble()})
           .toList();
 
       await ordersApi.createOrder(items: items);
